dcdc

+++ config/mail.php
@@ -51,7 +51,7 @@ return [
             'scheme' => env('MAIL_SCHEME'),
             'url' => env('MAIL_URL'),
             'host' => env('MAIL_HOST', '127.0.0.1'),
-            'port' => env('MAIL_PORT', 2525),
+            'port' => (int) env('MAIL_PORT', 2525),
             'username' => env('MAIL_USERNAME'),
             'password' => env('MAIL_PASSWORD'),
             'timeout' => null,
@@ -61,7 +61,7 @@ return [
         'transactional' => [
             'transport' => env('MAIL_TRANSACTIONAL_TRANSPORT', env('MAIL_MAILER', 'smtp')),
             'host' => env('MAIL_TRANSACTIONAL_HOST', env('MAIL_HOST')),
-            'port' => env('MAIL_TRANSACTIONAL_PORT', env('MAIL_PORT')),
+            'port' => (int) env('MAIL_TRANSACTIONAL_PORT', env('MAIL_PORT')),
             'username' => env('MAIL_TRANSACTIONAL_USERNAME', env('MAIL_USERNAME')),
             'password' => env('MAIL_TRANSACTIONAL_PASSWORD', env('MAIL_PASSWORD')),
             'timeout' => env('MAIL_TRANSACTIONAL_TIMEOUT', 30),
@@ -70,7 +70,7 @@ return [
         'marketing' => [
             'transport' => env('MAIL_MARKETING_TRANSPORT', env('MAIL_MAILER', 'smtp')),
             'host' => env('MAIL_MARKETING_HOST', env('MAIL_HOST')),
-            'port' => env('MAIL_MARKETING_PORT', env('MAIL_PORT')),
+            'port' => (int) env('MAIL_MARKETING_PORT', env('MAIL_PORT')),
             'username' => env('MAIL_MARKETING_USERNAME', env('MAIL_USERNAME')),
             'password' => env('MAIL_MARKETING_PASSWORD', env('MAIL_PASSWORD')),
             'timeout' => env('MAIL_MARKETING_TIMEOUT', 60),
